pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Farmer's Data Management System"
authors = ["Akash Damle <akash.damle@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-dotenv = "^1.4.2"
django-extensions = "^3.2.3"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
markdown = "^3.7"
pillow = "^11.0.0"
requests = "^2.32.3"
django-allauth = "^65.2.0"
dj-rest-auth = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
